backend/requirements.txt
api==0.0.7
asgiref==3.5.2
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
clarifai==2.6.2
clarifai-grpc==8.6.0
click==8.1.3
configparser==3.8.1
Django==4.0.5
django-jsonfield==1.4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
googleapis-common-protos==1.56.3
grpcio==1.47.0
idna==3.3
isort==5.10.1
jaraco.context==4.1.1
jsonschema==2.6.0
more-itertools==8.13.0
mypy-extensions==0.4.3
nose==1.3.7
pathspec==0.9.0
Pillow==9.1.1
pipfile==0.0.2
platformdirs==2.5.2
protobuf==3.19.3
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wolframalpha==5.0.0
xmltodict==0.13.0
